FIX Изменил структуру проекта
- добавил обработку 404 ошибки (страница не найдена)
