+++ app/templates/user/read_ad.html
@@ -54,7 +54,7 @@
                 <li class="page-item">
                     <a class="page-link {% if not pagination.has_prev %} disabled {% endif %}"
                        href="{% if pagination.has_prev %}
-                            {{ url_for('user_advertisement.view_ad_user', page=pagination.prev_num) }}
+                            {{ url_for('read_ad_user', page=pagination.prev_num) }}
                             {% else %}{% endif %}">Предыдущая
                     </a>
                 </li>
@@ -64,7 +64,7 @@
                         {% if pagination.page != number %}
                             <li class="page-item">
                                 <a class="page-link"
-                                   href="{{ url_for('user_advertisement.view_ad_user', page=number) }}">
+                                   href="{{ url_for('read_ad_user', page=number) }}">
                                     {{ number }}
                                 </a>
                             </li>
@@ -85,7 +85,7 @@
                 <li class="page-item">
                     <a class="page-link {% if not pagination.has_next %} disabled {% endif %}"
                        href="{% if pagination.has_next %}
-                            {{ url_for('user_advertisement.view_ad_user', page=pagination.next_num) }}
+                            {{ url_for('read_ad_user', page=pagination.next_num) }}
                             {% else %}{% endif %}">Следующая
                     </a>
                 </li>
